Revert "junhyeon vgg16 cat_Vs_dog"
